@@ -1,17 +1,18 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import { AuthContextProvider } from "./Contexts/AuthContextProvider";
-import { BrowserRouter } from "react-router-dom";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider as ReduxProvider } from "react-redux";
+import { store } from "./redux/store";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <AuthContextProvider>
+    <ReduxProvider store={store}>
+      <BrowserRouter>
         <App />
-      </AuthContextProvider>
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+      </BrowserRouter>
+    </ReduxProvider>
+  </React.StrictMode>
 );
